refactor: split animation.ts into math.ts, animation.ts, and dom.ts
- math.ts: pure math/physics helpers (randomBetween, randomStartPhase,
  randomDiagonalAngle, sineOffset, bounceAtBounds, randomYInBand)
- animation.ts: theme-aware animation primitives (createAnimationLoop,
  createBouncer, createGlider, createDrifter)
- dom.ts: DOM creation utilities (createGifPopup, createFallingItems,
  spawnAnimatedSprite)

Each file has a single responsibility and no cross-concerns.
Updated all theme imports and tests.
Updated LICENSE year to 2019-present.
