+++ src/themes/sharks.ts
@@ -1,6 +1,7 @@
 import type { Theme } from './types';
-import type { GifPopup } from '../utils/animation';
-import { createBouncer, createGifPopup } from '../utils/animation';
+import type { GifPopup } from '../utils/dom';
+import { createBouncer } from '../utils/animation';
+import { createGifPopup } from '../utils/dom';
 import { ScoreCounter } from '../utils/ScoreCounter';
 
 const THEME_ID = 'sharks';
